+++ src/containers/geo_location.jsx
@@ -10,6 +10,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
 import getLocation from '../actions/action_geolocation';
+import {fetchGeocodeReverse} from '../actions/action_fetch_geocode_reverse';
 import {fetchGeocodeWeather} from '../actions/action_fetch_geocode_weather_data';
 
 import GoogleMap from '../components/google_map';
@@ -28,8 +29,11 @@ class Location extends Component {
   }
 
   fetchLocal(){
+    const lat = this.props.location.coords.latitude;
+    const lon = this.props.location.coords.longitude;
 
-    this.props.fetchGeocodeWeather(this.props.location.coords.latitude, this.props.location.coords.longitude);
+    this.props.fetchGeocodeWeather(lat, lon);
+    this.props.fetchGeocodeReverse(lat, lon);
 
     var local = document.querySelector('#localWeather');
     var button = document.querySelector('#fetchWeather');
@@ -43,10 +47,11 @@ class Location extends Component {
 
   renderWhenReady(){
     
-    const weatherData = this.props.geoweather[0];
+    const weather = this.props.geoweather[0];
+    const cityData = this.props.geocodereverse[0];
     const lat = this.props.location.coords.latitude;
     const lon = this.props.location.coords.longitude;
-
+    
     if (lat <= 0 && lon <= 0) { 
 
       // console.log("Loading Geolocation...")   
@@ -63,7 +68,7 @@ class Location extends Component {
       // console.log(lon);
       // console.log(this.props.location);
 
-      if (weatherData == undefined) {
+      if (weather == undefined) {
         
         // console.log('weather data empty');
         var alerts;
@@ -74,27 +79,29 @@ class Location extends Component {
       }
 
       else {
-        
-        console.log('Geolocation weather data fetched! - ', weatherData);
-        
-        if (weatherData.alerts == undefined) {
+        var city = cityData.address_components[3].long_name;
+        var province = cityData.address_components[5].long_name;
+        var country = cityData.address_components[6].long_name; 
+        console.log('Geolocation weather data fetched! - ', weather);
+        console.log('Geolocation city data fetched! - ', cityData);
+        if (weather.alerts == undefined) {
           alerts = undefined;
         } else {
-          alerts = weatherData.alerts.map(alertData => alertData.description);
+          alerts = weather.alerts.map(alertData => alertData.description);
         }
         
         // console.log(alerts);
         
-        var summary = weatherData.hourly.summary;
+        var summary = weather.hourly.summary;
         // console.log(summary);
         
-        var temp = weatherData.hourly.data.map(temps => temps.temperature);
+        var temp = weather.hourly.data.map(temps => temps.temperature);
         // console.log(temp);
 
-        var humi = weatherData.hourly.data.map(humis => humis.humidity);
+        var humi = weather.hourly.data.map(humis => humis.humidity);
         // console.log(humi);
 
-        var wind = weatherData.hourly.data.map(winds => winds.windSpeed);
+        var wind = weather.hourly.data.map(winds => winds.windSpeed);
         // console.log(wind);
       }
 
@@ -103,16 +110,17 @@ class Location extends Component {
         <article id="geolocate" className="card animated fadeInUp" key={lat}>
           
           <div className="card-block">
-            <h4 className="card-title animated fadeInDown">Your current location is ...</h4>
+            <h4 className="card-title animated fadeInDown">Your approximate location is ... </h4>
           </div>
           
           <section id="geolocateMap">
             <div className="mapContainer animated fadeInDown">           
                 <GoogleMap zoom={16} lon={lon} lat={lat}/>
             </div> 
-            <button id="fetchWeather" className="btn btn-primary animated fadeInUp" onClick={this.fetchLocal}>Show local Forecast</button>
+            <button id="fetchWeather" className="btn btn-primary animated fadeInUp" onClick={this.fetchLocal}>Show your local hourly forecast</button>
           </section>  
           <div id="localWeather" className="hidden">
+            <h3 className="card-title animated fadeInDown">48 hour forecast for: {city}, {province}, {country}.</h3>
             <Charts key={lat} summary={summary} temp={temp} humi={humi} wind={wind} alerts={alerts}/>
             
           </div>  
@@ -148,12 +156,12 @@ class Location extends Component {
   }
 }
 
-function mapStateToProps({ geoweather, location}){
-  return {geoweather, location};
+function mapStateToProps({ geoweather, geocodereverse, location}){
+  return {geoweather, geocodereverse, location};
 }
 
 function mapDispatchToProps(dispatch) {
-  return bindActionCreators({fetchGeocodeWeather, getLocation}, dispatch);
+  return bindActionCreators({fetchGeocodeWeather, fetchGeocodeReverse, getLocation}, dispatch);
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(Location);